# improve disk queue performance by coalescing writes