@cmacknz 2022-06-15
LGTM, I'll let @faec do the approving here. The test failure is fixed by https://github.com/elastic/beats/pull/31943